+++ hub/src/containers/curriculum/Detail.js
