+++ hub/src/containers/curriculum/Metainfo.js
